+++ packages/cosmo/src/schema/auth.ts
@@ -1,0 +1,10 @@
+import { Schema } from "effect";
+
+export const RefreshTokenResultSchema = Schema.Struct({
+  refreshToken: Schema.String,
+  accessToken: Schema.String,
+});
+
+export const RefreshResponseSchema = Schema.Struct({
+  credentials: RefreshTokenResultSchema,
+});
