+++ packages/cosmo/src/schema/common.ts
@@ -1,0 +1,4 @@
+import { Schema } from "effect";
+import { validArtists } from "../types/common.js";
+
+export const ValidArtistSchema = Schema.Literal(...validArtists);
