+++ packages/cosmo/src/types/metadata.ts
@@ -1,39 +1,8 @@
-export type CosmoObjektMetadataV1 = {
-  name: string;
-  description: string;
-  image: string;
-  background_color: string;
-  objekt: {
-    collectionId: string;
-    season: string;
-    member: string;
-    collectionNo: string;
-    class: string;
-    artists: string[];
-    thumbnailImage: string;
-    frontImage: string;
-    backImage: string;
-    accentColor: string;
-    backgroundColor: string;
-    textColor: string;
-    comoAmount: number;
-    tokenId: string;
-    objektNo: number;
-    tokenAddress: string;
-    transferable: boolean;
-  };
-};
+import type { MetadataV1Schema, MetadataV3Schema } from "../schema/metadata.js";
 
-export type CosmoObjektMetadataV3 = {
-  name: string;
-  description: string;
-  image: string;
-  background_color: string;
-  attributes: {
-    trait_type: string;
-    value: string;
-  }[];
-};
+export type CosmoObjektMetadataV1 = typeof MetadataV1Schema.Type;
+
+export type CosmoObjektMetadataV3 = typeof MetadataV3Schema.Type;
 
 /**
  * Attempt to convert v3 metadata to v1 metadata.
